requirements.txt
fastapi == 0.109.2
python-dotenv == 1.0.1
openai == 1.12.0
uvicorn[standard]==0.27.1
pytest==8.0.2
